preproc.awk: fix for arm64/aarch64 SBSA: number of CPUs
This fixes THREADS for an arm64 machine I was using, while also still
working on other systems.

The output of "numactl --hardware" on that system yielded this:

    node 1 cpus:
    note 2 cpus:
    ...etc...

This causes preproc.awk to only find one cpu.

Instead of using an awk loop to count CPUs, just feed the answer into
the script via a new NCPUS awk variable. nproc(1) does a fine job of
counting CPUs.

Add a check for nproc(1)'s existence, in order to ease first time use.

Signed-off-by: John Hubbard <jhubbard@nvidia.com>