add compile_commands.json generation, via bear(1), if present
This is still rudimentary, because the system generates multiple
binaries from the same source file, but with different defines being
passed in. That confounds clangd's ability to figure out the values of
these defines, even with compile_commands.json covering all of the
binaries.

Further refactoring is needed in order to browse the source code sanely.
Probably best to generate different source files for each build, so that
these can be browsed (and even hacked up for local experiments).

Signed-off-by: John Hubbard <jhubbard@nvidia.com>